# Conflicts:
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/data/message/Message.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/data/message/MessageContent.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/data/message/PersistMessageUseCase.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/data/message/ProtoContentMapper.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/message/MessageScope.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/message/SessionResetSender.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/sessionreset/ResetSessionUseCase.kt
#	persistence/src/commonMain/kotlin/com/wire/kalium/persistence/dao/message/MessageDAOImpl.kt

+++ persistence/src/androidMain/kotlin/com/wire/kalium/persistence/dao/message/KaliumPager.kt
@@ -9,8 +9,8 @@ import kotlinx.coroutines.flow.Flow
  * Exposes a [pagingDataFlow] that can be used in Android UI components to display paginated data.
  */
 class KaliumPager<EntityType : Any>(
-    private val pager: Pager<Long, EntityType>,
-    internal val pagingSource: PagingSource<Long, EntityType>
+    private val pager: Pager<Int, EntityType>,
+    internal val pagingSource: PagingSource<Int, EntityType>
 ) {
     val pagingDataFlow: Flow<PagingData<EntityType>>
         get() = pager.flow
